requirements.txt
simpy           # simulazione processi
matplotlib      # grafici
python-dotenv   # gestione variabili di ambiente
pydantic        # validazione dei dati
asyncio         # operazioni asincrone
grpcio          # gRPC framework
grpcio-tools    # gRPC generazione codice
types-protobuf  # tipizzazione dei messaggi protobuf
pyyaml			# gestione file yaml
types-PyYAML	# tipizzazione libreria pyyaml

# === TESTING ===
pytest                    # Testing framework
pytest-cov                # Coverage reports
pytest-asyncio			  # Async

# === CODE QUALITY ===
black                      # Code formatting
isort                      # Import sorting
flake8                     # Linting
mypy                       # Type checking
bandit                     # Security linting